+++ Migrations/KsiegarniaOnlineContextModelSnapshot.cs
@@ -74,8 +74,8 @@ namespace KsiegarniaOnline.Migrations
                     b.Property<int>("NumberOfPages")
                         .HasColumnType("int");
 
-                    b.Property<double>("Price")
-                        .HasColumnType("float");
+                    b.Property<int>("Price")
+                        .HasColumnType("int");
 
                     b.Property<int>("PublishHouseID")
                         .HasColumnType("int");
